move db update code to database.js, update mutex use add a unique constrain to species table fix regions
