+++ js/database.js
@@ -162,4 +162,99 @@ function closeDatabase(db) {
     }
   });
 }
-export { sqlite3, closeDatabase, checkpoint, Mutex };
+
+
+async function upgrade_to_v1(diskDB, dbMutex){
+  let t0 = Date.now()
+  try{
+    await dbMutex.lock();
+    await diskDB.runAsync("PRAGMA foreign_keys=OFF");
+    await diskDB.runAsync("BEGIN");
+    //1.10.x update
+    await diskDB.runAsync("CREATE INDEX IF NOT EXISTS idx_species_sname ON species(sname)");
+    await diskDB.runAsync("CREATE INDEX IF NOT EXISTS idx_species_cname ON species(cname)");
+    const fileColumns = (await diskDB.allAsync("PRAGMA table_info(files)")).map(row => row.name);
+    if (!fileColumns.includes("archiveName")){
+      await diskDB.runAsync("ALTER TABLE files ADD COLUMN archiveName TEXT");  
+    }
+    if (!fileColumns.includes("metadata")){
+      await diskDB.runAsync("ALTER TABLE files ADD COLUMN metadata TEXT");  
+    }
+
+    await diskDB.runAsync("CREATE TABLE IF NOT EXISTS tags(id INTEGER PRIMARY KEY, name TEXT NOT NULL, UNIQUE(name))");
+    await diskDB.runAsync("INSERT INTO tags VALUES(0, 'Nocmig'), (1, 'Local')");
+    await diskDB.runAsync("ALTER TABLE records ADD COLUMN tagID INTEGER");
+    await diskDB.runAsync("UPDATE records SET tagID = 0 WHERE label = 'Nocmig'");
+    await diskDB.runAsync("UPDATE records SET tagID = 1 WHERE label = 'Local'");
+    await diskDB.runAsync("ALTER TABLE records DROP COLUMN label");
+    // Change label names to labelIDs
+    await diskDB.runAsync("ALTER TABLE records ADD COLUMN reviewed INTEGER");
+
+    await diskDB.runAsync(`CREATE TABLE records_temp( dateTime INTEGER, position INTEGER, fileID INTEGER, speciesID INTEGER, confidence INTEGER, 
+      comment  TEXT, end INTEGER, callCount INTEGER, isDaylight INTEGER, reviewed INTEGER, tagID INTEGER,
+      UNIQUE (dateTime, fileID, speciesID), 
+      CONSTRAINT fk_files FOREIGN KEY (fileID) REFERENCES files(id) ON DELETE CASCADE,
+      CONSTRAINT fk_species FOREIGN KEY (speciesID) REFERENCES species(id),
+      CONSTRAINT fk_tags FOREIGN KEY (tagID) REFERENCES tags(id) ON DELETE SET NULL)`);
+    await diskDB.runAsync("INSERT INTO records_temp SELECT * from records");
+    await diskDB.runAsync("DROP TABLE records");
+    await diskDB.runAsync("ALTER TABLE records_temp RENAME TO records");
+    // Add old files table update
+    await diskDB.runAsync(`
+      CREATE TABLE files_new (
+          id INTEGER PRIMARY KEY, 
+          name TEXT NOT NULL, 
+          duration REAL,
+          filestart INTEGER, 
+          locationID INTEGER, 
+          archiveName TEXT, 
+          metadata TEXT, 
+          UNIQUE (name),
+          CONSTRAINT fk_locations FOREIGN KEY (locationID) REFERENCES locations(id) ON DELETE SET NULL
+      )`);
+    await diskDB.runAsync("INSERT INTO files_new SELECT * FROM files");
+    await diskDB.runAsync("DROP TABLE files");
+    await diskDB.runAsync("ALTER TABLE files_new RENAME TO files");
+    await diskDB.runAsync("DROP TABLE IF EXISTS db_upgrade");
+    await diskDB.runAsync("END");
+    await diskDB.runAsync("PRAGMA foreign_keys=ON");
+    await diskDB.runAsync("PRAGMA integrity_check");
+    await diskDB.runAsync("PRAGMA foreign_key_check");
+    await diskDB.runAsync("PRAGMA user_version = 1");
+    console.info("Migrated tags and added 'reviewed' column to ", diskDB.filename)
+  } catch (e) {
+    console.error("Error adding column and updating version", e.message, e)
+    await diskDB.runAsync("ROLLBACK");
+  } finally{
+    await checkpoint(diskDB)
+    dbMutex.unlock();
+    console.info(`DB migration took ${Date.now() - t0}ms`)
+  }
+}
+
+async function upgrade_to_v2(diskDB, dbMutex){
+  let t0 = Date.now();
+  try{
+    await dbMutex.lock();
+    await diskDB.runAsync("PRAGMA foreign_keys=OFF");
+    await diskDB.runAsync("BEGIN");
+    await diskDB.runAsync("CREATE TABLE species_new(id INTEGER PRIMARY KEY, sname TEXT NOT NULL, cname TEXT NOT NULL, UNIQUE(cname, sname))");
+    await diskDB.runAsync("INSERT INTO species_new SELECT * FROM species");
+    await diskDB.runAsync("DROP TABLE species");
+    await diskDB.runAsync("ALTER TABLE species_new RENAME TO species");
+    await diskDB.runAsync("END");
+    await diskDB.runAsync("PRAGMA foreign_keys=ON");
+    await diskDB.runAsync("PRAGMA integrity_check");
+    await diskDB.runAsync("PRAGMA foreign_key_check");
+    await diskDB.runAsync("PRAGMA user_version = 2");
+    console.info(`Adding species unique constraint took ${Date.now() - t0}ms`)
+  } catch (e) {
+    console.error("Error adding unique constraint ", e.message, e)
+    await diskDB.runAsync("ROLLBACK");
+  } finally{
+    await checkpoint(diskDB)
+    dbMutex.unlock();
+  }
+}
+
+export { sqlite3, closeDatabase, checkpoint, upgrade_to_v1, upgrade_to_v2, Mutex };
